+++ priv/templates/coh.install/models/coherence/invitation.ex
@@ -4,12 +4,10 @@ defmodule <%= base %>.Coherence.Invitation do
   """
   use Ecto.Schema
   import Ecto.Changeset
-
   <%= if use_binary_id? do %>
   @primary_key {:id, :binary_id, autogenerate: true}
   @foreign_key_type :binary_id
-  <% end %>
-
+  <% else %><% end %>
   schema "invitations" do
     field(:name, :string)
     field(:email, :string)
@@ -24,7 +22,7 @@ defmodule <%= base %>.Coherence.Invitation do
   If no params are provided, an invalid changeset is returned
   with no validation performed.
   """
-  @spec changeset(Ecto.Schema.t, Map.t) :: Ecto.Changeset.t
+  @spec changeset(Ecto.Schema.t(), Map.t()) :: Ecto.Changeset.t()
   def changeset(model, params \\ %{}) do
     model
     |> cast(params, ~w(name email token))
@@ -36,7 +34,7 @@ defmodule <%= base %>.Coherence.Invitation do
   @doc """
   Creates a changeset for a new schema
   """
-  @spec new_changeset(Map.t) :: Ecto.Changeset.t
+  @spec new_changeset(Map.t()) :: Ecto.Changeset.t()
   def new_changeset(params \\ %{}) do
     changeset(%__MODULE__{}, params)
   end